requirements.txt
# requirements.txt
streamlit>=1.37.0
python-dotenv>=1.0.1
strands-agents
strands-agents-tools
pypdf
scikit-learn